requirements.txt
boto3
botocore
click
cloudpickle
dask
datashader
debugpy
elasticsearch_dsl
fuzzymatcher
geopandas
geopy
holoviews
interchange
joblib
matplotlib
mlflow
neo4j
nest_asyncio
nltk
numpy
ogb
optuna
packaging
pandas
paramiko
plotly
pmdarima
psutil
py2neo
pyarrow
python_dateutil
python_dotenv
pytorch_lightning
pytz
PyYAML
requests
requests_futures
scikit_learn
scp
smart_open
statsmodels
thefuzz
tqdm
transformers
umap_learn
urllib3
python-dotenv
umap_learn
crossref-commons
semanticscholar
dgl
ray
scipy
torch